+++ lib/src/widgets/widget.dart
@@ -4,11 +4,13 @@ import 'package:flutter/services.dart';
 import 'package:url_launcher/url_launcher.dart' as url_launcher;
 
 import '../platform_utils/platform_utils.dart';
+import './preview_container.dart';
 import './properties_container.dart';
 import './stories_list.dart';
 import './icon.dart';
 import './helpers.dart';
 import '../story.dart';
+import './intructions_dialog.dart';
 
 class _DashbookDualTheme {
   final ThemeData light;
@@ -149,23 +151,11 @@ class _DashbookState extends State<Dashbook> {
               child: Stack(
                 children: [
                   if (_currentChapter != null)
-                    Positioned.fill(
-                      child: Container(
-                        decoration: BoxDecoration(
-                          border: !widget.usePreviewSafeArea
-                              ? null
-                              : Border(
-                                  left: BorderSide(
-                                      color: Theme.of(context).cardColor,
-                                      width: iconSize(context) * 2),
-                                  right: BorderSide(
-                                      color: Theme.of(context).cardColor,
-                                      width: iconSize(context) * 2),
-                                ),
-                        ),
-                        child: chapterWidget,
-                      ),
+                    PreviewContainer(
                       key: Key(_currentChapter!.id),
+                      child: chapterWidget!,
+                      usePreviewSafeArea: widget.usePreviewSafeArea,
+                      isPropertiesOpen: _currentView == CurrentView.PROPERTIES,
                     ),
                   Positioned(
                     right: 10,
@@ -180,6 +170,19 @@ class _DashbookState extends State<Dashbook> {
                             onClick: () => setState(
                                 () => _currentView = CurrentView.PROPERTIES),
                           ),
+                        if (_currentChapter?.info != null)
+                          DashbookIcon(
+                              tooltip: 'Instructions',
+                              icon: Icons.info,
+                              onClick: () {
+                                showDialog(
+                                    context: context,
+                                    builder: (_) {
+                                      return InstructionsDialog(
+                                        instructions: _currentChapter!.info!,
+                                      );
+                                    });
+                              }),
                         if (_currentChapter?.codeLink != null)
                           DashbookIcon(
                             tooltip: 'See code',