+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [next]
+## [0.1.2]
  - Adding control property to all properties
  - Large screens preview container share the space with the properties contaner for better visualization
  - Adding info property to chapters